requirements.txt
pandas~=1.4.4
pvlib~=0.9.1
dash==2.6.1
plotly~=5.10.0
psycopg2~=2.9.3
python-dotenv~=0.21.0